+++ blocc/member/member_at.cpp
@@ -67,7 +67,7 @@ Value& MemberATExpression::value(Context& ctx) const
   {
     Collection * rv = val.collection();
     Integer p = *a0.integer();
-    if (p >= 0 && p < rv->size())
+    if (p >= 0 && size_t(p) < rv->size())
       return rv->at((unsigned)p).to_lvalue(val.lvalue());
     throw RuntimeError(EXC_RT_INDEX_RANGE_S, a0.toString().c_str());
   }
@@ -78,7 +78,7 @@ Value& MemberATExpression::value(Context& ctx) const
   {
     Literal * rv = val.literal();
     Integer p = *a0.integer();
-    if (p >= 0 && p < rv->size())
+    if (p >= 0 && size_t(p) < rv->size())
       return ctx.allocate(Value(Integer((unsigned char)rv->at((unsigned) p))));
     throw RuntimeError(EXC_RT_INDEX_RANGE_S, a0.toString().c_str());
   }
@@ -86,7 +86,7 @@ Value& MemberATExpression::value(Context& ctx) const
   {
     TabChar * rv = val.tabchar();
     Integer p = *a0.integer();
-    if (p >= 0 && p < rv->size())
+    if (p >= 0 && size_t(p) < rv->size())
       return ctx.allocate(Value(Integer((unsigned char)rv->at((unsigned) p))));
     throw RuntimeError(EXC_RT_INDEX_RANGE_S, a0.toString().c_str());
   }
